add jenkins-hijack-agent
This is a departure of the previous style of implementing scripts.
Rather than having scripts that implement a bunch of bash functions my
new preference is to have them be separate scripts, this way they can be
implemented in different languages.

There is another repository for my ~/bin directory but I chose to keep
that private for now.

Now to the fun part: this script will take a list of URLs that point to
Jenkins servers and find ones with a specific label (hardcoded as cloud,
but should probably be customizable in the future). After it finds one
that is idle, it update the labels and start a new shell to allow for
ssh, etc. After the shell is terminated it will change the label back to
the original value, thus releasing the agent.